fix: failed migration

+++ database/migrations/Invoice/2023_08_16_155117_delate_invoice_product_list.php
@@ -10,6 +10,8 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::dropIfExists('InvoiceProductList');
+        if (Schema::hasTable('InvoiceProductList')) {
+            Schema::dropIfExists('InvoiceProductList');
+        }
     }
 };
